Moved EntityService into its own folder
