+++ lib/services/entity_service/entity_service.dart
